Guard stale skill execution reports
