+++ agents/Aevatar.GAgents.Scheduled/protos/skill_definition.proto
@@ -35,6 +35,7 @@ message SkillDefinitionState {
   optional int32 max_tokens = 18;
   optional int32 max_tool_rounds = 19;
   optional int32 max_history_messages = 20;
+  int64 config_revision = 21;
 
   reserved 8, 10, 11, 12;
 }
@@ -93,6 +94,7 @@ message SkillDefinitionExecutionDispatchFailedEvent {
 message ReportSkillExecutionCompletedCommand {
   string execution_id = 1;
   google.protobuf.Timestamp completed_at = 2;
+  optional int64 definition_config_revision = 3;
 }
 
 message ReportSkillExecutionFailedCommand {
@@ -100,11 +102,13 @@ message ReportSkillExecutionFailedCommand {
   google.protobuf.Timestamp failed_at = 2;
   string error = 3;
   int32 retry_attempt = 4;
+  optional int64 definition_config_revision = 5;
 }
 
 message SkillDefinitionExecutionCompletedEvent {
   string execution_id = 1;
   google.protobuf.Timestamp completed_at = 2;
+  optional int64 definition_config_revision = 3;
 }
 
 message SkillDefinitionExecutionFailedEvent {
@@ -112,6 +116,7 @@ message SkillDefinitionExecutionFailedEvent {
   google.protobuf.Timestamp failed_at = 2;
   string error = 3;
   int32 retry_attempt = 4;
+  optional int64 definition_config_revision = 5;
 }
 
 message DisableSkillDefinitionCommand {
