+++ test/Aevatar.GAgents.ChannelRuntime.Tests/SkillExecutionGAgentTests.cs
@@ -1,0 +1,854 @@
+using System.Net;
+using System.Reflection;
+using System.Text;
+using System.Text.Json;
+using Aevatar.AI.Abstractions.LLMProviders;
+using Aevatar.AI.ToolProviders.NyxId;
+using Aevatar.Foundation.Abstractions;
+using Aevatar.Foundation.Abstractions.Persistence;
+using Aevatar.Foundation.Core;
+using Aevatar.Foundation.Core.EventSourcing;
+using FluentAssertions;
+using Google.Protobuf;
+using Microsoft.Extensions.DependencyInjection;
+using NSubstitute;
+using Aevatar.GAgents.Scheduled;
+
+namespace Aevatar.GAgents.ChannelRuntime.Tests;
+
+public sealed class SkillExecutionGAgentTests : IAsyncLifetime
+{
+    private InMemoryEventStore _store = null!;
+    private ServiceProvider _serviceProvider = null!;
+
+    public Task InitializeAsync()
+    {
+        _store = new InMemoryEventStore();
+        _serviceProvider = BuildServiceProvider(_store);
+        return Task.CompletedTask;
+    }
+
+    public Task DisposeAsync()
+    {
+        _serviceProvider.Dispose();
+        return Task.CompletedTask;
+    }
+
+    [Fact]
+    public async Task SendOutputAsync_ShouldUseTypedReceiveTarget_WhenLarkReceiveIdIsPopulated()
+    {
+        // Initialize with typed fields set (the shape AgentBuilderTool now writes for p2p flows).
+        // Even though the legacy ConversationId is an `oc_*` chat id (which Lark would also accept
+        // with chat_id), the typed open_id target should be sent verbatim — this is what fixes the
+        // production 400 where the relay's ConversationId fell through to ou_*.
+        var outbound = new SkillRunnerOutboundConfig
+        {
+            ConversationId = "oc_chat_legacy",
+            NyxProviderSlug = "api-lark-bot",
+            NyxApiKey = "nyx-api-key",
+            LarkReceiveId = "ou_user_1",
+            LarkReceiveIdType = "open_id",
+        };
+        var agent = await CreateActivatedExecutionAgent("exec-typed-target", outbound);
+
+        var handler = new RecordingHandler("""{"code":0,"msg":"success","data":{"message_id":"om_1"}}""");
+        AttachNyxIdApiClient(agent, handler);
+
+        await InvokeSendOutputAsync(agent, "scheduled report body");
+
+        handler.LastRequest.Should().NotBeNull();
+        handler.LastRequest!.RequestUri!.ToString()
+            .Should().Be("https://nyx.example.com/api/v1/proxy/s/api-lark-bot/open-apis/im/v1/messages?receive_id_type=open_id");
+        using var body = JsonDocument.Parse(handler.LastBody!);
+        body.RootElement.GetProperty("receive_id").GetString().Should().Be("ou_user_1");
+        body.RootElement.GetProperty("msg_type").GetString().Should().Be("text");
+    }
+
+    [Fact]
+    public async Task SendOutputAsync_ShouldFallBackToConversationIdPrefixInference_ForLegacyState()
+    {
+        // Backward compatibility: state persisted before the typed lark_receive_id fields existed
+        // still resolves through the prefix heuristic on ConversationId. The send still succeeds
+        // (no exception); the sender emits a Debug breadcrumb that is not visible to xUnit.
+        var outbound = new SkillRunnerOutboundConfig
+        {
+            ConversationId = "ou_legacy_user",
+            NyxProviderSlug = "api-lark-bot",
+            NyxApiKey = "nyx-api-key",
+        };
+        var agent = await CreateActivatedExecutionAgent("exec-legacy-fallback", outbound);
+
+        var handler = new RecordingHandler("""{"code":0,"msg":"success"}""");
+        AttachNyxIdApiClient(agent, handler);
+
+        await InvokeSendOutputAsync(agent, "legacy report body");
+
+        handler.LastRequest!.RequestUri!.ToString()
+            .Should().Be("https://nyx.example.com/api/v1/proxy/s/api-lark-bot/open-apis/im/v1/messages?receive_id_type=open_id");
+        using var body = JsonDocument.Parse(handler.LastBody!);
+        body.RootElement.GetProperty("receive_id").GetString().Should().Be("ou_legacy_user");
+    }
+
+    [Fact]
+    public async Task SendOutputAsync_ShouldThrow_WhenLarkBusinessCodeIsNonZero()
+    {
+        // Lark reports business errors as HTTP 200 with `code != 0`. Ignoring the response would
+        // let HandleTriggerAsync persist SkillRunnerExecutionCompletedEvent on a silent failure.
+        var outbound = new SkillRunnerOutboundConfig
+        {
+            ConversationId = "oc_chat_1",
+            NyxProviderSlug = "api-lark-bot",
+            NyxApiKey = "nyx-api-key",
+            LarkReceiveId = "ou_user_1",
+            LarkReceiveIdType = "open_id",
+        };
+        var agent = await CreateActivatedExecutionAgent("exec-lark-biz-err", outbound);
+
+        var handler = new RecordingHandler("""{"code":230002,"msg":"invalid receive_id"}""");
+        AttachNyxIdApiClient(agent, handler);
+
+        Func<Task> act = () => InvokeSendOutputAsync(agent, "report");
+
+        var assertion = await act.Should().ThrowAsync<InvalidOperationException>();
+        assertion.WithMessage("*code=230002*");
+        assertion.WithMessage("*invalid receive_id*");
+    }
+
+    [Fact]
+    public async Task SendOutputAsync_ShouldThrow_WhenNyxProxyEnvelopeReportsError()
+    {
+        // HTTP non-2xx from NyxID gets packaged into a Nyx envelope that ProxyRequestAsync returns
+        // verbatim. Ignoring it would mask transport / auth failures.
+        var outbound = new SkillRunnerOutboundConfig
+        {
+            ConversationId = "oc_chat_1",
+            NyxProviderSlug = "api-lark-bot",
+            NyxApiKey = "nyx-api-key",
+            LarkReceiveId = "ou_user_1",
+            LarkReceiveIdType = "open_id",
+        };
+        var agent = await CreateActivatedExecutionAgent("exec-nyx-envelope-err", outbound);
+
+        var handler = new RecordingHandler("""{"error":true,"message":"upstream timeout"}""");
+        AttachNyxIdApiClient(agent, handler);
+
+        Func<Task> act = () => InvokeSendOutputAsync(agent, "report");
+
+        await act.Should().ThrowAsync<InvalidOperationException>()
+            .WithMessage("*upstream timeout*");
+    }
+
+    [Fact]
+    public async Task SendOutputAsync_ShouldIncludeRecreateHint_When_LarkRejectsAsCrossAppOpenId()
+    {
+        // PR #409 review (pulls/409#review-4175198266): after this fix new agents capture
+        // union_id, but agents created before the fix still have `LarkReceiveIdType=open_id`
+        // pinned to a relay-app-scoped `ou_*`. Their next scheduled run hits Lark
+        // `99992361 open_id cross app` and the user sees the bare error in `/agent-status`'s
+        // `last_error` with no clue what to do. Surface explicit "delete and recreate" guidance
+        // so the failure becomes self-documenting.
+        var outbound = new SkillRunnerOutboundConfig
+        {
+            ConversationId = "oc_chat_1",
+            NyxProviderSlug = "api-lark-bot",
+            NyxApiKey = "nyx-api-key",
+            LarkReceiveId = "ou_relay_app_user_1",
+            LarkReceiveIdType = "open_id",
+        };
+        var agent = await CreateActivatedExecutionAgent("exec-cross-app-openid", outbound);
+
+        var handler = new RecordingHandler(
+            """{"code":99992361,"msg":"open_id cross app","error":{"message":"Refer to the documentation"}}""");
+        AttachNyxIdApiClient(agent, handler);
+
+        Func<Task> act = () => InvokeSendOutputAsync(agent, "report");
+
+        var assertion = await act.Should().ThrowAsync<InvalidOperationException>();
+        assertion.WithMessage("*code=99992361*");
+        assertion.WithMessage("*open_id cross app*");
+        // The hint must be actionable enough that the user can recover without reading source.
+        assertion.WithMessage("*before cross-app union_id ingress existed*");
+        assertion.WithMessage("*/agents*");
+        assertion.WithMessage("*Delete*");
+        assertion.WithMessage("*/daily*");
+    }
+
+    [Fact]
+    public async Task SendOutputAsync_ShouldRetryWithFallback_When_PrimaryRejectedAsBotNotInChat_ViaHttp400Envelope()
+    {
+        // Reviewer (PR #412 r3141700469): production failures arrive through
+        // `NyxIdApiClient.SendAsync` as an HTTP-400 Nyx envelope:
+        // `{"error": true, "status": 400, "body": "{\"code\":230002,...}"}`. The previous
+        // `LarkProxyResponse.TryGetError` returned true for that shape but left
+        // `larkCode=null` because it didn't parse the nested `body`, so the BotNotInChat
+        // retry branch never fired in the actual production path. Pin the wrapped envelope
+        // shape end-to-end.
+        var outbound = new SkillRunnerOutboundConfig
+        {
+            ConversationId = "oc_dm_chat_1",
+            NyxProviderSlug = "api-lark-bot",
+            NyxApiKey = "nyx-api-key",
+            LarkReceiveId = "oc_dm_chat_1",
+            LarkReceiveIdType = "chat_id",
+            LarkReceiveIdFallback = "on_user_1",
+            LarkReceiveIdTypeFallback = "union_id",
+        };
+        var agent = await CreateActivatedExecutionAgent("exec-fallback-http400", outbound);
+
+        // First (primary) attempt: NyxIdApiClient.SendAsync HTTP-400 envelope wrapping Lark
+        // 230002. Second (fallback) attempt: clean success.
+        var handler = new SequencedHandler(
+            """{"error": true, "status": 400, "body": "{\"code\":230002,\"msg\":\"Bot is not in the chat\"}"}""",
+            """{"code":0,"msg":"success"}""");
+        AttachNyxIdApiClient(agent, handler);
+
+        await InvokeSendOutputAsync(agent, "report");
+
+        handler.Requests.Should().HaveCount(2);
+        handler.Requests[0].RequestUri!.Query.Should().Contain("receive_id_type=chat_id");
+        handler.Requests[1].RequestUri!.Query.Should().Contain("receive_id_type=union_id");
+        handler.Bodies[1].Should().Contain("\"receive_id\":\"on_user_1\"");
+    }
+
+    [Fact]
+    public async Task SendOutputAsync_ShouldThrowCrossTenantHint_When_LarkCodeNestedInHttp400Body()
+    {
+        // Same envelope shape as the production /daily failure log: NyxID wraps the Lark
+        // 99992364 as a string body inside an HTTP-400 Nyx envelope. The cross-tenant
+        // recreate-the-agent hint (PR #412) only fires when the parser surfaces the nested
+        // Lark code; previously it never did. Pin both the recovery hint and the nested-body
+        // unwrap together.
+        var outbound = new SkillRunnerOutboundConfig
+        {
+            ConversationId = "oc_dm_chat_1",
+            NyxProviderSlug = "api-lark-bot",
+            NyxApiKey = "nyx-api-key",
+            LarkReceiveId = "on_relay_tenant_user_1",
+            LarkReceiveIdType = "union_id",
+        };
+        var agent = await CreateActivatedExecutionAgent("exec-cross-tenant-http400", outbound);
+
+        var handler = new RecordingHandler(
+            """{"error": true, "status": 400, "body": "{\"code\":99992364,\"msg\":\"user id cross tenant\"}"}""");
+        AttachNyxIdApiClient(agent, handler);
+
+        Func<Task> act = () => InvokeSendOutputAsync(agent, "report");
+
+        var assertion = await act.Should().ThrowAsync<InvalidOperationException>();
+        assertion.WithMessage("*99992364*");
+        assertion.WithMessage("*different tenant*");
+        assertion.WithMessage("*/agents*");
+        assertion.WithMessage("*Delete*");
+        assertion.WithMessage("*/daily*");
+    }
+
+    [Fact]
+    public async Task SendOutputAsync_ShouldRetryWithFallback_When_PrimaryRejectedAsBotNotInChat()
+    {
+        // Reviewer concern (codex-bot, P1, PR #412): chat_id-first regresses cross-app
+        // same-tenant deployments where the outbound app is not a member of the inbound DM
+        // chat — Lark returns `230002 bot not in chat` for chat_id-typed sends. Captured the
+        // union_id at create time as a fallback; assert the runtime retries once with the
+        // fallback typed pair when the primary attempt fails with 230002, and that the retry
+        // body uses the fallback `receive_id` / `receive_id_type`.
+        var outbound = new SkillRunnerOutboundConfig
+        {
+            ConversationId = "oc_dm_chat_1",
+            NyxProviderSlug = "api-lark-bot",
+            NyxApiKey = "nyx-api-key",
+            LarkReceiveId = "oc_dm_chat_1",
+            LarkReceiveIdType = "chat_id",
+            LarkReceiveIdFallback = "on_user_1",
+            LarkReceiveIdTypeFallback = "union_id",
+        };
+        var agent = await CreateActivatedExecutionAgent("exec-fallback-bot-not-in-chat", outbound);
+
+        var handler = new SequencedHandler(
+            """{"code":230002,"msg":"Bot is not in the chat"}""",
+            """{"code":0,"msg":"success"}""");
+        AttachNyxIdApiClient(agent, handler);
+
+        await InvokeSendOutputAsync(agent, "report");
+
+        handler.Requests.Should().HaveCount(2);
+        handler.Requests[0].RequestUri!.Query.Should().Contain("receive_id_type=chat_id");
+        handler.Bodies[0].Should().Contain("\"receive_id\":\"oc_dm_chat_1\"");
+        handler.Requests[1].RequestUri!.Query.Should().Contain("receive_id_type=union_id");
+        handler.Bodies[1].Should().Contain("\"receive_id\":\"on_user_1\"");
+    }
+
+    [Fact]
+    public async Task SendOutputAsync_ShouldNotRetry_When_PrimaryRejectedWithDifferentLarkCode()
+    {
+        // Only `230002 bot not in chat` triggers the fallback retry. Other Lark codes (e.g.
+        // 99992364 cross_tenant) propagate immediately so the user sees the actionable
+        // recovery hint for the actual failure mode rather than a misleading retry.
+        var outbound = new SkillRunnerOutboundConfig
+        {
+            ConversationId = "oc_dm_chat_1",
+            NyxProviderSlug = "api-lark-bot",
+            NyxApiKey = "nyx-api-key",
+            LarkReceiveId = "oc_dm_chat_1",
+            LarkReceiveIdType = "chat_id",
+            LarkReceiveIdFallback = "on_user_1",
+            LarkReceiveIdTypeFallback = "union_id",
+        };
+        var agent = await CreateActivatedExecutionAgent("exec-no-retry-diff-code", outbound);
+
+        var handler = new SequencedHandler(
+            """{"code":99992364,"msg":"user id cross tenant"}""");
+        AttachNyxIdApiClient(agent, handler);
+
+        Func<Task> act = () => InvokeSendOutputAsync(agent, "report");
+
+        await act.Should().ThrowAsync<InvalidOperationException>()
+            .WithMessage("*99992364*");
+        handler.Requests.Should().ContainSingle("only 230002 should trigger the fallback retry");
+    }
+
+    [Fact]
+    public async Task SendOutputAsync_ShouldIncludeRecreateHint_When_LarkRejectsAsCrossTenantUserId()
+    {
+        // Production failure mode after PR #409 switched p2p to union_id: NyxID's relay-side
+        // ingress and `s/api-lark-bot` proxy turned out to be in different Lark tenants, so even
+        // union_id is rejected. This PR pivots to chat_id-first; the cross_tenant error code is
+        // surfaced with the same recreate guidance so legacy agents (still pinned to union_id)
+        // give users a way to recover without reading source.
+        var outbound = new SkillRunnerOutboundConfig
+        {
+            ConversationId = "oc_chat_1",
+            NyxProviderSlug = "api-lark-bot",
+            NyxApiKey = "nyx-api-key",
+            LarkReceiveId = "on_relay_tenant_user_1",
+            LarkReceiveIdType = "union_id",
+        };
+        var agent = await CreateActivatedExecutionAgent("exec-cross-tenant-userid", outbound);
+
+        var handler = new RecordingHandler(
+            """{"code":99992364,"msg":"user id cross tenant","error":{"log_id":"L1"}}""");
+        AttachNyxIdApiClient(agent, handler);
+
+        Func<Task> act = () => InvokeSendOutputAsync(agent, "report");
+
+        var assertion = await act.Should().ThrowAsync<InvalidOperationException>();
+        assertion.WithMessage("*code=99992364*");
+        assertion.WithMessage("*user id cross tenant*");
+        assertion.WithMessage("*different tenant*");
+        assertion.WithMessage("*chat_id-preferred*");
+        assertion.WithMessage("*/agents*");
+        assertion.WithMessage("*Delete*");
+        assertion.WithMessage("*/daily*");
+    }
+
+    [Fact]
+    public async Task TrySendFailureAsync_ShouldUseFailureNotificationSlug_WhenSetAndDistinctFromPrimary()
+    {
+        // Issue #423 §C: when a primary outbound delivery has just been rejected (e.g. cross-
+        // tenant 99992364), retrying the failure notification through the SAME slug also
+        // fails. The fix routes failure-notifications through the inbound channel-bot's slug
+        // captured at agent-create time — by definition reachable, since the user just
+        // messaged it. This test pins that the routing actually changes the proxy slug in
+        // the outbound URL while the receive_id, body, and api key stay identical.
+        var outbound = new SkillRunnerOutboundConfig
+        {
+            ConversationId = "oc_dm_chat_1",
+            NyxProviderSlug = "api-lark-bot",
+            NyxApiKey = "nyx-api-key",
+            LarkReceiveId = "ou_user_1",
+            LarkReceiveIdType = "open_id",
+            FailureNotificationProviderSlug = "api-lark-bot-channel-loning",
+        };
+        var agent = await CreateActivatedExecutionAgent("exec-failure-slug", outbound);
+
+        var handler = new RecordingHandler("""{"code":0,"msg":"success","data":{"message_id":"om_failure"}}""");
+        AttachNyxIdApiClient(agent, handler);
+
+        await InvokeTrySendFailureAsync(agent, "Lark message delivery rejected (code=99992364): user id cross tenant.");
+
+        handler.LastRequest.Should().NotBeNull();
+        handler.LastRequest!.RequestUri!.ToString()
+            .Should().Be("https://nyx.example.com/api/v1/proxy/s/api-lark-bot-channel-loning/open-apis/im/v1/messages?receive_id_type=open_id");
+        using var body = JsonDocument.Parse(handler.LastBody!);
+        body.RootElement.GetProperty("receive_id").GetString().Should().Be("ou_user_1");
+        // The failure-notification message itself should carry the original error so the user
+        // sees the actionable hint in chat, not just an unsubscribe ping.
+        var contentJson = body.RootElement.GetProperty("content").GetString();
+        contentJson.Should().Contain("99992364");
+        contentJson.Should().Contain("Skill execution failed");
+    }
+
+    [Fact]
+    public async Task TrySendFailureAsync_ShouldFallBackToPrimarySlug_WhenFailureNotificationSlugRejects()
+    {
+        // Defense-in-depth: the failure-notification slug MAY have been valid at create time
+        // (in user-services) but become inactive (token revoked, bot uninstalled, etc.) by
+        // the time the agent fires. If its send rejects, we still try the primary slug as
+        // a last-resort attempt — better than the user seeing nothing.
+        var outbound = new SkillRunnerOutboundConfig
+        {
+            ConversationId = "oc_dm_chat_1",
+            NyxProviderSlug = "api-lark-bot",
+            NyxApiKey = "nyx-api-key",
+            LarkReceiveId = "ou_user_1",
+            LarkReceiveIdType = "open_id",
+            FailureNotificationProviderSlug = "api-lark-bot-channel-revoked",
+        };
+        var agent = await CreateActivatedExecutionAgent("exec-failure-slug-rejects", outbound);
+
+        var handler = new SequencedHandler(
+            // Failure-notification slug rejects with a Nyx envelope error (e.g. 401 on the proxy)
+            """{"error":true,"message":"upstream auth failed"}""",
+            // Primary slug succeeds — best-effort recovery.
+            """{"code":0,"msg":"success","data":{"message_id":"om_primary_failure"}}""");
+        AttachNyxIdApiClient(agent, handler);
+
+        await InvokeTrySendFailureAsync(agent, "report rejected at primary");
+
+        handler.Requests.Should().HaveCount(2);
+        handler.Requests[0].RequestUri!.ToString()
+            .Should().Contain("/proxy/s/api-lark-bot-channel-revoked/");
+        handler.Requests[1].RequestUri!.ToString()
+            .Should().Contain("/proxy/s/api-lark-bot/");
+    }
+
+    [Fact]
+    public async Task TrySendFailureAsync_ShouldSkipFailureSlug_WhenEqualToPrimary()
+    {
+        // No-op fallback: if the inbound slug equals the primary slug there is no recovery
+        // benefit — same proxy = same rejection mode. AgentBuilderTool leaves the field
+        // empty in this case, but pin the runtime guard too so a future
+        // mis-capture doesn't pay double-POST cost just to fail twice.
+        var outbound = new SkillRunnerOutboundConfig
+        {
+            ConversationId = "oc_dm_chat_1",
+            NyxProviderSlug = "api-lark-bot",
+            NyxApiKey = "nyx-api-key",
+            LarkReceiveId = "ou_user_1",
+            LarkReceiveIdType = "open_id",
+            FailureNotificationProviderSlug = "api-lark-bot",
+        };
+        var agent = await CreateActivatedExecutionAgent("exec-failure-slug-equal", outbound);
+
+        var handler = new RecordingHandler("""{"code":0,"msg":"success"}""");
+        AttachNyxIdApiClient(agent, handler);
+
+        await InvokeTrySendFailureAsync(agent, "primary failed");
+
+        // Exactly one POST — no double-attempt against the same slug.
+        handler.LastRequest.Should().NotBeNull();
+        handler.LastRequest!.RequestUri!.ToString()
+            .Should().Contain("/proxy/s/api-lark-bot/");
+    }
+
+    [Fact]
+    public async Task TrySendFailureAsync_ShouldUsePrimary_WhenFailureNotificationSlugIsEmpty()
+    {
+        // Backwards compat: agents created before #423 §C have an empty
+        // FailureNotificationProviderSlug. The runtime must transparently fall back to the
+        // existing single-attempt behavior — this test is the regression guard against the
+        // failure-notification fallback ever introducing a hidden dependency on the new
+        // field being populated.
+        var outbound = new SkillRunnerOutboundConfig
+        {
+            ConversationId = "oc_dm_chat_1",
+            NyxProviderSlug = "api-lark-bot",
+            NyxApiKey = "nyx-api-key",
+            LarkReceiveId = "ou_user_1",
+            LarkReceiveIdType = "open_id",
+            // FailureNotificationProviderSlug intentionally not set (legacy state shape).
+        };
+        var agent = await CreateActivatedExecutionAgent("exec-failure-slug-empty", outbound);
+
+        var handler = new RecordingHandler("""{"code":0,"msg":"success"}""");
+        AttachNyxIdApiClient(agent, handler);
+
+        await InvokeTrySendFailureAsync(agent, "primary failed");
+
+        handler.LastRequest.Should().NotBeNull();
+        handler.LastRequest!.RequestUri!.ToString()
+            .Should().Contain("/proxy/s/api-lark-bot/");
+    }
+
+    [Fact]
+    public async Task TrySendFailureAsync_ShouldSwallow_WhenBothSlugsReject()
+    {
+        // Final guarantee: TrySendFailureAsync MUST NOT throw, ever. HandleTriggerAsync is
+        // already in the failure-event-persist path; an exception here would mask the
+        // SkillRunnerExecutionFailedEvent persist (which surfaces last_error in
+        // /agent-status, the one path users have to recover regardless of Lark visibility).
+        var outbound = new SkillRunnerOutboundConfig
+        {
+            ConversationId = "oc_dm_chat_1",
+            NyxProviderSlug = "api-lark-bot",
+            NyxApiKey = "nyx-api-key",
+            LarkReceiveId = "ou_user_1",
+            LarkReceiveIdType = "open_id",
+            FailureNotificationProviderSlug = "api-lark-bot-channel-loning",
+        };
+        var agent = await CreateActivatedExecutionAgent("exec-failure-slug-both-reject", outbound);
+
+        var handler = new SequencedHandler(
+            """{"error":true,"message":"failure-slug down"}""",
+            """{"code":99992364,"msg":"user id cross tenant"}""");
+        AttachNyxIdApiClient(agent, handler);
+
+        // Should complete (not throw) even though both attempts fail.
+        Func<Task> act = () => InvokeTrySendFailureAsync(agent, "both broken");
+
+        await act.Should().NotThrowAsync();
+        handler.Requests.Should().HaveCount(2);
+    }
+
+    [Fact]
+    public async Task BuildExecutionMetadata_ShouldPinOwnerLlmConfigOverrides_WhenSourceReturnsConfig()
+    {
+        // Regression for the "/daily failed: Provider 'openai' not connected" report:
+        // skill runners must honor the bot owner's pre-configured model + NyxID route + tool
+        // cap — same shape ChannelLlmReplyInboxRuntime applies for nyxid-chat. Without it,
+        // every scheduled run falls through to NyxIdLLMProvider's compile-time `gpt-5.4` +
+        // gateway default, which the gateway routes to OpenAI and 400s for bot owners who
+        // wired a custom NyxID service like `chrono-llm` at `/api/v1/proxy/s/chrono-llm`.
+        var source = new StubOwnerLlmConfigSource(new OwnerLlmConfig(
+            DefaultModel: "gpt-5.5",
+            PreferredLlmRoute: "/api/v1/proxy/s/chrono-llm",
+            MaxToolRounds: 7));
+
+        var agent = await CreateActivatedExecutionAgent(
+            "skill-runner-userconfig",
+            CreateInitializeCommand().OutboundConfig,
+            source);
+
+        var metadata = await InvokeBuildExecutionMetadataAsync(agent);
+
+        metadata[LLMRequestMetadataKeys.ModelOverride].Should().Be("gpt-5.5");
+        metadata[LLMRequestMetadataKeys.NyxIdRoutePreference].Should().Be("/api/v1/proxy/s/chrono-llm");
+        metadata[LLMRequestMetadataKeys.MaxToolRoundsOverride].Should().Be("7");
+        source.RequestedScopeIds.Should().ContainSingle().Which.Should().Be("scope-1");
+    }
+
+    [Fact]
+    public async Task BuildExecutionMetadata_ShouldOmitOverrides_WhenOwnerLlmConfigSourceIsAbsent()
+    {
+        // No host wiring (e.g. tests that don't compose Studio + the bridge): valid metadata
+        // still comes out, no override keys leak, NyxIdLLMProvider falls through to its
+        // compile-time defaults.
+        var agent = await CreateActivatedExecutionAgent(
+            "skill-runner-userconfig-no-source",
+            CreateInitializeCommand().OutboundConfig);
+
+        var metadata = await InvokeBuildExecutionMetadataAsync(agent);
+
+        metadata.Should().NotContainKey(LLMRequestMetadataKeys.ModelOverride);
+        metadata.Should().NotContainKey(LLMRequestMetadataKeys.NyxIdRoutePreference);
+        metadata.Should().NotContainKey(LLMRequestMetadataKeys.MaxToolRoundsOverride);
+        metadata[LLMRequestMetadataKeys.NyxIdAccessToken].Should().Be("nyx-api-key");
+    }
+
+    [Fact]
+    public async Task BuildExecutionMetadata_ShouldOmitOverrides_WhenOwnerLlmConfigFieldsAreEmpty()
+    {
+        // Bot owners who haven't saved any LLM preference get OwnerLlmConfig.Empty (or empty
+        // strings via the host adapter). The applier must NOT pin empty values onto metadata,
+        // because NyxIdLLMProvider treats a non-empty NyxIdRoutePreference of "" as a relative
+        // path against the authority and produces an invalid URL.
+        var source = new StubOwnerLlmConfigSource(OwnerLlmConfig.Empty);
+
+        var agent = await CreateActivatedExecutionAgent(
+            "skill-runner-userconfig-empty",
+            CreateInitializeCommand().OutboundConfig,
+            source);
+
+        var metadata = await InvokeBuildExecutionMetadataAsync(agent);
+
+        metadata.Should().NotContainKey(LLMRequestMetadataKeys.ModelOverride);
+        metadata.Should().NotContainKey(LLMRequestMetadataKeys.NyxIdRoutePreference);
+        metadata.Should().NotContainKey(LLMRequestMetadataKeys.MaxToolRoundsOverride);
+    }
+
+    [Fact]
+    public async Task BuildExecutionMetadata_ShouldFallBackQuietly_WhenOwnerLlmConfigSourceThrows()
+    {
+        // The source can throw on transient projection failures. The agent's execution turn
+        // must still proceed with provider defaults — the applier catches and logs the
+        // failure, never bubbles it up to the trigger handler.
+        var source = new ThrowingOwnerLlmConfigSource();
+
+        var agent = await CreateActivatedExecutionAgent(
+            "skill-runner-userconfig-throws",
+            CreateInitializeCommand().OutboundConfig,
+            source);
+
+        var metadata = await InvokeBuildExecutionMetadataAsync(agent);
+
+        metadata.Should().NotContainKey(LLMRequestMetadataKeys.ModelOverride);
+        metadata.Should().NotContainKey(LLMRequestMetadataKeys.NyxIdRoutePreference);
+    }
+
+    private static async Task<IReadOnlyDictionary<string, string>> InvokeBuildExecutionMetadataAsync(
+        SkillExecutionGAgent agent)
+    {
+        var method = typeof(SkillExecutionGAgent).GetMethod(
+            "BuildExecutionMetadataAsync",
+            BindingFlags.Instance | BindingFlags.NonPublic);
+        method.Should().NotBeNull();
+        var task = (Task<IReadOnlyDictionary<string, string>>)method!.Invoke(agent, [CancellationToken.None])!;
+        return await task;
+    }
+
+    internal sealed class StubOwnerLlmConfigSource(OwnerLlmConfig config) : IOwnerLlmConfigSource
+    {
+        public List<string> RequestedScopeIds { get; } = new();
+
+        public Task<OwnerLlmConfig> GetForScopeAsync(string scopeId, CancellationToken ct = default)
+        {
+            RequestedScopeIds.Add(scopeId);
+            return Task.FromResult(config);
+        }
+    }
+
+    internal sealed class ThrowingOwnerLlmConfigSource : IOwnerLlmConfigSource
+    {
+        public Task<OwnerLlmConfig> GetForScopeAsync(string scopeId, CancellationToken ct = default) =>
+            throw new InvalidOperationException("projection unavailable");
+    }
+
+    private static void AttachNyxIdApiClient(SkillExecutionGAgent agent, HttpMessageHandler handler)
+    {
+        var client = new NyxIdApiClient(
+            new NyxIdToolOptions { BaseUrl = "https://nyx.example.com" },
+            new HttpClient(handler) { BaseAddress = new Uri("https://nyx.example.com") });
+        var field = typeof(SkillExecutionGAgent).GetField(
+            "_nyxIdApiClient",
+            BindingFlags.Instance | BindingFlags.NonPublic);
+        field.Should().NotBeNull();
+        field!.SetValue(agent, client);
+    }
+
+    private static Task InvokeSendOutputAsync(SkillExecutionGAgent agent, string output)
+    {
+        var method = typeof(SkillExecutionGAgent).GetMethod(
+            "SendOutputAsync",
+            BindingFlags.Instance | BindingFlags.NonPublic,
+            binder: null,
+            types: [typeof(string), typeof(CancellationToken)],
+            modifiers: null);
+        method.Should().NotBeNull();
+        return (Task)method!.Invoke(agent, [output, CancellationToken.None])!;
+    }
+
+    private static Task InvokeTrySendFailureAsync(SkillExecutionGAgent agent, string error)
+    {
+        var method = typeof(SkillExecutionGAgent).GetMethod(
+            "TrySendFailureAsync",
+            BindingFlags.Instance | BindingFlags.NonPublic);
+        method.Should().NotBeNull();
+        return (Task)method!.Invoke(agent, [error, CancellationToken.None])!;
+    }
+
+    private sealed class RecordingHandler(string responseBody) : HttpMessageHandler
+    {
+        public HttpRequestMessage? LastRequest { get; private set; }
+        public string? LastBody { get; private set; }
+
+        protected override async Task<HttpResponseMessage> SendAsync(HttpRequestMessage request, CancellationToken cancellationToken)
+        {
+            LastRequest = request;
+            LastBody = request.Content == null
+                ? null
+                : await request.Content.ReadAsStringAsync(cancellationToken);
+            return new HttpResponseMessage(HttpStatusCode.OK)
+            {
+                Content = new StringContent(responseBody, Encoding.UTF8, "application/json"),
+            };
+        }
+    }
+
+    /// <summary>
+    /// Returns a different response per request in the order given. Used to simulate the
+    /// `bot not in chat` rejection on the primary attempt followed by a successful fallback
+    /// retry.
+    /// </summary>
+    private sealed class SequencedHandler : HttpMessageHandler
+    {
+        private readonly Queue<string> _responses;
+        public List<HttpRequestMessage> Requests { get; } = new();
+        public List<string?> Bodies { get; } = new();
+
+        public SequencedHandler(params string[] responses)
+        {
+            _responses = new Queue<string>(responses);
+        }
+
+        protected override async Task<HttpResponseMessage> SendAsync(HttpRequestMessage request, CancellationToken cancellationToken)
+        {
+            Requests.Add(request);
+            Bodies.Add(request.Content == null ? null : await request.Content.ReadAsStringAsync(cancellationToken));
+            var body = _responses.Count > 0 ? _responses.Dequeue() : """{"code":0,"msg":"success"}""";
+            return new HttpResponseMessage(HttpStatusCode.OK)
+            {
+                Content = new StringContent(body, Encoding.UTF8, "application/json"),
+            };
+        }
+    }
+
+    private SkillExecutionGAgent CreateExecutionAgent(string actorId, IOwnerLlmConfigSource? ownerLlmConfigSource = null)
+    {
+        var agent = new SkillExecutionGAgent(ownerLlmConfigSource: ownerLlmConfigSource)
+        {
+            Services = _serviceProvider,
+            EventSourcingBehaviorFactory =
+                _serviceProvider.GetRequiredService<IEventSourcingBehaviorFactory<SkillExecutionState>>(),
+        };
+        AssignActorId(agent, actorId);
+        return agent;
+    }
+
+    /// <summary>
+    /// Creates a <see cref="SkillExecutionGAgent"/>, seeds a
+    /// <see cref="SkillExecutionStartedEvent"/> into the event store so the agent's state
+    /// is fully populated with the given outbound config after activation (via event replay),
+    /// then activates it. This avoids calling <c>HandleStartAsync</c> which would trigger
+    /// the actual LLM execution path.
+    /// </summary>
+    private async Task<SkillExecutionGAgent> CreateActivatedExecutionAgent(
+        string actorId,
+        SkillRunnerOutboundConfig outboundConfig,
+        IOwnerLlmConfigSource? ownerLlmConfigSource = null)
+    {
+        var startedEvent = new SkillExecutionStartedEvent
+        {
+            DefinitionId = "def-test",
+            StartedAt = Google.Protobuf.WellKnownTypes.Timestamp.FromDateTimeOffset(DateTimeOffset.UtcNow),
+            Reason = "test",
+            SkillContent = "You are a test skill.",
+            ExecutionPrompt = "Run the test.",
+            OutboundConfig = outboundConfig.Clone(),
+            ScopeId = "scope-1",
+            ProviderName = SkillDefinitionDefaults.DefaultProviderName,
+        };
+
+        var stateEvent = new StateEvent
+        {
+            EventId = Guid.NewGuid().ToString("N"),
+            Timestamp = Google.Protobuf.WellKnownTypes.Timestamp.FromDateTimeOffset(DateTimeOffset.UtcNow),
+            Version = 1,
+            EventType = SkillExecutionStartedEvent.Descriptor.FullName,
+            EventData = Google.Protobuf.WellKnownTypes.Any.Pack(startedEvent),
+            AgentId = actorId,
+        };
+        await _store.AppendAsync(actorId, [stateEvent], 0);
+
+        var agent = CreateExecutionAgent(actorId, ownerLlmConfigSource);
+        await agent.ActivateAsync();
+        return agent;
+    }
+
+    private static ServiceProvider BuildServiceProvider(
+        IEventStore eventStore,
+        Action<IServiceCollection>? configure = null)
+    {
+        var services = new ServiceCollection();
+        services.AddSingleton(eventStore);
+        services.AddSingleton<EventSourcingRuntimeOptions>();
+        services.AddTransient(
+            typeof(IEventSourcingBehaviorFactory<>),
+            typeof(DefaultEventSourcingBehaviorFactory<>));
+        configure?.Invoke(services);
+        return services.BuildServiceProvider();
+    }
+
+    private static InitializeSkillDefinitionCommand CreateInitializeCommand() => new()
+    {
+        SkillName = "daily_report",
+        TemplateName = "daily_report",
+        SkillContent = "You are a daily report runner.",
+        ExecutionPrompt = "Run the report.",
+        ScheduleCron = string.Empty,
+        ScheduleTimezone = SkillDefinitionDefaults.DefaultTimezone,
+        Enabled = true,
+        ScopeId = "scope-1",
+        ProviderName = SkillDefinitionDefaults.DefaultProviderName,
+        OutboundConfig = new SkillRunnerOutboundConfig
+        {
+            ConversationId = "oc_chat_1",
+            NyxProviderSlug = "api-lark-bot",
+            NyxApiKey = "nyx-api-key",
+        },
+    };
+
+    private static void AssignActorId(GAgentBase agent, string actorId)
+    {
+        var setIdMethod = typeof(GAgentBase).GetMethod(
+            "SetId",
+            BindingFlags.Instance | BindingFlags.NonPublic);
+        setIdMethod.Should().NotBeNull();
+        setIdMethod!.Invoke(agent, [actorId]);
+    }
+
+    private sealed class InMemoryEventStore : IEventStore
+    {
+        private readonly Dictionary<string, List<StateEvent>> _events = new(StringComparer.Ordinal);
+
+        public Task<EventStoreCommitResult> AppendAsync(
+            string agentId,
+            IEnumerable<StateEvent> events,
+            long expectedVersion,
+            CancellationToken ct = default)
+        {
+            ct.ThrowIfCancellationRequested();
+            if (!_events.TryGetValue(agentId, out var stream))
+            {
+                stream = [];
+                _events[agentId] = stream;
+            }
+
+            var currentVersion = stream.Count == 0 ? 0 : stream[^1].Version;
+            if (currentVersion != expectedVersion)
+                throw new InvalidOperationException(
+                    $"Optimistic concurrency conflict: expected {expectedVersion}, actual {currentVersion}");
+
+            var appended = events.Select(x => x.Clone()).ToList();
+            stream.AddRange(appended);
+            return Task.FromResult(new EventStoreCommitResult
+            {
+                AgentId = agentId,
+                LatestVersion = stream[^1].Version,
+                CommittedEvents = { appended.Select(x => x.Clone()) },
+            });
+        }
+
+        public Task<IReadOnlyList<StateEvent>> GetEventsAsync(
+            string agentId,
+            long? fromVersion = null,
+            CancellationToken ct = default)
+        {
+            ct.ThrowIfCancellationRequested();
+            if (!_events.TryGetValue(agentId, out var stream))
+                return Task.FromResult<IReadOnlyList<StateEvent>>([]);
+
+            IReadOnlyList<StateEvent> result = fromVersion.HasValue
+                ? stream.Where(x => x.Version > fromVersion.Value).Select(x => x.Clone()).ToList()
+                : stream.Select(x => x.Clone()).ToList();
+            return Task.FromResult(result);
+        }
+
+        public Task<long> GetVersionAsync(string agentId, CancellationToken ct = default)
+        {
+            ct.ThrowIfCancellationRequested();
+            if (!_events.TryGetValue(agentId, out var stream) || stream.Count == 0)
+                return Task.FromResult(0L);
+            return Task.FromResult(stream[^1].Version);
+        }
+
+        public Task<long> DeleteEventsUpToAsync(string agentId, long toVersion, CancellationToken ct = default)
+        {
+            ct.ThrowIfCancellationRequested();
+            if (toVersion <= 0 || !_events.TryGetValue(agentId, out var stream))
+                return Task.FromResult(0L);
+
+            var before = stream.Count;
+            stream.RemoveAll(x => x.Version <= toVersion);
+            return Task.FromResult((long)(before - stream.Count));
+        }
+    }
+}
